+++ CHANGELOG.md
@@ -4,6 +4,18 @@
 ### Added
 - Forward/Right arrow key will fill in hint text (#327)
 - Values returned by bait hooks will be passed to the `throw` caller
+- `display` property to completion groups entries to style completion entries when type is `list`.
+example:
+```lua
+local cg = {
+	items = {
+		'list item 1',
+		['--command-flag-here'] = {'this does a thing', '--the-flag-alias'},
+		['--styled-command-flag-here'] = {'this does a thing', '--the-flag-alias', display = lunacolors.blue '--styled-command-flag-here'}
+	},
+	type = 'list'
+}
+```
 
 ## [2.3.4] - 2024-12-28
 ### Fixed