+++ CHANGELOG.md
@@ -5,11 +5,13 @@
 - Forward/Right arrow key will fill in hint text (#327)
 ### Changed
 - Documentation for Lunacolors has been improved, with more information added.
+- Values returned by bait hooks will be passed to the `throw` caller
 
 ## [2.3.4] - 2024-12-28
 ### Fixed
 - Skip over file and prevent panic if info cannot be retrieved during file completion (due to permission error or anything else)
 - Apply environment variables properly after 2.3 shell interpreter changes
+- hilbish.sink.readAll() function now reads data that doesn't end in a newline
 
 ## [2.3.3] - 2024-11-04
 ### Fixed